cryptonews aggregator python project/crytpo_news_agg/pyproject.toml
[tool.poetry]
name = "crytpo_news_agg"
version = "0.1.0"
description = ""
authors = ["saif mujahed abdullah hayel <prof.msaif@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
psycopg2 = "^2.9.1"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.4"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
